[FIRRTL] Add Verif/LTL intrinsics
Add intrinsics to the FIRRTL dialect that lower directly to operations
in the Verif and LTL dialects. This allows FIRRTL inputs to express
temporal assert, assume, and cover operations without requiring native
support for them in the FIRRTL spec.

This commit does the following:

- Add `firrtl.int.ltl.*` and `firrtl.int.verif.*` operations to
  `FIRRTLExpressions.td` and `FIRRTLStatements.td`
- Add intrinsic lowering support to `LowerIntrinsics`
- Add visitor support
- In LowerToHW, map the intrinsics to the corresponding `verif.*` and
  `ltl.*` operations.
- Since there are no opaque user-defined FIRRTL types yet, the
  intrinsics operate on `UInt<1>` where the LTL ops would operate on
  concrete `!ltl.sequence` and `!ltl.property` types. The LowerToHW
  change also includes a final `fixupLTLOps` step to ensure these types
  are propagate properly and don't leak into `i1` logic.
- Add tests for `LowerIntrinsics`, `LowerToHW`, and an end-to-end test
  with firtool to check for FIR-to-SVA output.

Now that the number of intrinsics starts to grow, we probably want to
refactor the `LowerIntrinsics` pass a bit and create something like a
pattern/intrinsic set and a few helpers to match for ports and
parameters, with error reporting in one unified place. Future work.