[Python] Add CMake for MacOS C++17 compatibility in wheel build.
We are using C++17, and a recent addition to the Python bindings for
the OM dialect relies on a C++17 feature. For this to compile with
pybind11, we need to ensure we set the -mmacosx-version-min flag to at
least version 10.14, which is where support for these C++17 features
became available. By default, CMake was detecting an older version, so
explicitly set this when building the wheel.