[SV][ETC] Fix perf bug removing from a SetVector of slice results.
On one design (-Odebug), ETC was taking 600s / 1100s total,
after this change only takes ~20s.

Just drop the `remove` entirely, as it'll avoid re-slicing through
the same operation again (minor) and more to the point AFAICT we add the
"rootOp" removed here back to the result sets in the end anyway.