[Arc] Preserve extern modules as internal inputs/outputs (#5069)
Preserve instances of external modules in the arcilator pipeline rather
than removing them and replacing their outputs with zero. The inputs of
the extmodule act like outputs of the overall circuit, and the outputs
of the extmodule act like inputs of the overall circuit. During state
lowering, create write-only states for extmodule inputs (like taps) that
just capture whatever the circuit applies to the extmodule, and create
read-only states for extmodule outputs that the circuit only reads. This
allows the user of the software model to treat extmodule ports like
regular ports of the circuit and read/write the corresponding states.

Remove the extmodule stubbing in `StripSV`, since the default behavior
of the read/write-only states is to be reset to zero. So if the user of
the model does not write the extmodule outputs to any value, they will
remain at zero.