[Pipeline] Remove LI notion from pipelines
Removes the notion that `pipeline` pipelines can be either latency insensitive or sensitive.
This is done because:
1. only latency sensitive lowerings are supported at the moment
2. there seems to be no clear way to infer backpressure through the pipeline without this being explicitly defined in the IR.

The latter is obviously the more grave concern. If/until this is eventually found out, we can add back latency-insensitivity. But until then, we should not advertise a feature that has no clear path of support.