[Seq] Add clock gate conversion to extern module
Add the `ExternalizeClockGate` pass which converts all `seq.clock_gate`
ops in a design into instances of a `hw.module.extern`. The name and
ports of the extern module can be configured through pass options. In
case the module has no test enable port, the conversion adds a `comb.or`
to fold the enable and test enable operands of the `seq.clock_gate` into
a single enable port.