[FIRRTL] Add property assign op instead of using connect
This adds a new statement to FIRRTL to assign a property to a proptery
sink.  We stop allowing properties to be connected with a ConnectOp.
This new operation enforces that the parent operation is a FModuleOp,
and that it has static single connects, which is a good limitation to
keep from using this operation inside a WhenOp.