[FIRRTL][InferWidths] Skip over property types
No property types have inferrable widths, so we need to teach
InferWidths to gracefully ignore them.  Prior to this change, the only
non-base type in FIRRTL was RefType, which was convertable to its
underlying base type through `getBaseType(FIRRTLType)`.  This function
will now return null when the type is not convertible to a base type,
which includes property types. Inferwidths now null-checks the result to
skip over this situation.