[FIRRTL][InferWidths] Skip over property types (#5381)
No property types have inferable widths, so we need to teach
InferWidths to gracefully ignore them. InferWidths assumes that every
type is convertible to a base type by getBaseType(FIRRTLType). This
function will now return null when the type is not convertible to a base
type, which includes property types. Inferwidths now checks to see if
the conversion to a base type was successful.