[FIRRTL] Change `getBaseTypeOrNull` to `getBaseTypeIfPresent`
This changes `getBaseTypeOrNull` to accept a null type field and renames
the function to `getBaseTypeIfPresent`.  This aligns with upstream by
making this family of function (`OrNull`) accept a null argument, and
uses the newer naming to which makes this property more clear. There are
no users of this function outside of FIRRTLUtils.h.