[FIRRTL][InferWidths] Tweak debug printing to show names and more bits.
I've found that 0xFFF is not enough to always produce unique strings
for constraints I'm debugging, bump a bit from 4096 to 65536 for now
(still keeping it short but more bits to lower chance of collision).

Also add the result of `getFieldName()` to the debug printing for
each variable to make it easier, especially for blockarguments,
to match each to its origin.