[Pipeline] add 'go' signal to pipeline
This commit adds an explicit 'go' signal to the pipeline abstraction.
This signal is used to indicate when the pipeline should start. The
signal will propagate through each stage as the stage valid signal.
As a result of this, each stage now has a `s#_valid : i1` signal as the
last value in its block argument list. This value can be used within
each stage for any operations which requires access to the pipeline
control circuitry.
Given that pipeline stage validity is now explicit within the block
arguments of a stage, a `pipeline.stage` operation no longer has an
`enable` signal. As a small 'bugfix' here, a `seq.clock_gate` is emitted
to gate the pipeline stage separating registers on the pipeline stage
valid signal.