[Dedup] Parallelize hash calculation, NFCI
This commit parallelizes hash calculation. This reduces Dedup execution time
by 40% in a large design, which reduces the entire excution time by 7~8%.

This PR first changes the type of structual hash into <sha256, ArrayAttr>.
The first element is structual hash, the second element is a list of module names
that are referred by instances in the module. By doing so we can lazily combine
hashes of modules names replaced by dedup through `dedupMap`.