[Transforms] Make CSE extensible and implement CSE that ignores namehints
This commit modifies upstream CSE so that downstream users can use
their own OperationInfo and OperationEquivalence, and implement
`NamehintsInsensitiveCSE` which is a CSE pass that ignores namehints.
Since the current upstream CSE is hard to customize, I duplicated the relevant
code to lib/Transforms/CSE.cpp and will work on upstreaming patches afterward.

Changes that need to be upstreamed:

* CSE
Currently CSE is not implemented as a library, so this commit separates
CSE pass into `struct CSEImpl<OperationInfo>` where `OperationInfo` is
`DenseMapInfo` used for hashing and equivalence checking in CSE. Otherwise,
there is no functional difference to the original implementation.

* OperationEquivalence
`OperationEquivalence::computeHash` and `OperationEquivalence::isEquivalentTo`
are modified to take an additional callback to provide a way to customize CSE
(in our use case, functions that ignore `sv.namehint` are passed).