[LowerFIRRTLToHW] Create less wires with instance ports
When we lower input ports on instances, if there are zero or multiple
writers to the input port, we create a wire and use that as the operand
to the new instance operation.  There was a change to this code to use
`hw::WireOps`, which have their input value as an operand.  Since
`hw::WireOp`'s cannot have multiple writers, the code is now "the last
writer wins".  The wire's input is updated  when processing connect-like
operations - we will replace the previous writer to the wire.

We already have a helper in this code to gracefully handle this
situation called `createBackedge`.  A backedge is a placeholder for a
value which will be determined later.  We can use a backedge as an
operand to the lowered instance op, and when we find the connect
operation driving a value to the old instance port, replace the backedge
with the driven value.

The backedge helper in this pass gracefully handles undriven backedges
and combinational loops of backedges; if a backedge has not been
replaced by the end of the module body, it will be replaced with an
undriven wire.

An assert had to be removed which prevented a backedge from being
updated multiple times. Regular ~Connect ops were updated to support
backedges, instead of just strict connects.

This change is NFC in the FIRRTL pipeline, where the corner cases of
combinational loops and undriven ports are illegal. This better supports
the two illegal cases of undriven inputs and combinational loops. In the
future we should replace all usages of `createTmpHWWireOp` with the
backedge helpers.