[FIRRTL][IMDCE] Add test for dead output ports, NFC
This adds a test to show the behaviour that when a dead module
instantiates a child module, and uses its output port, that the dead
module should not demand the child's port and keep it alive.
Technically FIRRTL does not allow dead modules in the circuit, but this
pass already supports this.