[LowerFIRRTLToHW] Use backedges over temporary wires, NFC (#5513)
This changes LowerFIRRTLToHW to use backedges instead of temporary wire
operations. The temporary wires were guaranteed to be optimized away.
Now we slightly more gracefully handle the, technically illegal, cases
of these wires having no value driven to them.  This change is NFC in
the FIRRTL pipeline where these invariants have already been verified.
We can remove the infrastructure for creating these temporary wires, as
it is not used anymore.

Foreign typed ports can never be considered an InOut port, so they can
be handled by the code path which checks if the port is an input port,
which is handled identically.