[InferWidths] Remove expensive walks, NFC (#5523)
In the preprocess there seems to be an unintended use of `walk`.  Currently we walk the entire circuit three times in the pre-process. A walk in the post-order doesn't return early even when `skip()` was returned (see mlir/IR/Visitor.h). This commit just replaces walks with `getOps`. With this commit InferWidth should visit the entire IR only once.