[FIRRTL] Remove UninferredWidthCastOp for now.
Unclear what this models/means, and instead of doubling down
on workarounds lets drop this until we sort out
what this is and how it should work.

Fixes #5391.
(dropping the introduced back-propagation behavior)

This partially reverts commit 63d94ea003dcd0ed6c2f3da5580728569f2b346d.

Keep the parts of #5047 that aren't related to this op,
such as use of emitConnect throughout code.

StrictConnect still keeps the restriction regarding passive types
added in that commit, but now requires the operands to be sized as well.

We now need logic to emit ConstCast before a "connect",
which is added (under existing/modified tests).