[Support/ESI] Add generic port conversion utility (#5533)
This commit takes the ESI `ChannelRewriter` and generalizes it to provide a generic utility for performing port conversions on a `hw::HWMutableModuleLike` op.

It is intended to be a generic utility that can facilitate replacement of a given module in- or output to an arbitrary set of new inputs and outputs (i.e. 1 port -> N in, M out ports). Typical usecase is where an input (/output) of a module represents has some higher-level type that will be implemented by a set of lower-level typed in- and outputs ports + supporting operations within a module.

It also attempts to do so in an optimal way, by e.g. being able to collect multiple port modifications of a module, and perform them all at once.

A change wrt. the `ChannelRewriter` is that struct flattening has been removed. This is both to make the lowering more generic (hard to support the notion of "if this ESI channel type contains a struct, and we need to flatten the struct, ...", and is redundant, due to the `hw-flatten-io` pass.