[OM] Extract interfaces for ClassLike ops and ClassFieldLike ops.
This adds interfaces around the existing ClassOp and ClassFieldOp. The
interfaces are just abstracting over generated methods for the
ClassOp. For the ClassFieldOp, the interface just has one method to
get a type, which comes from the field's value.

By adding these interfaces and using them for parsing, printing,
verifying, etc., it will be simple to add new kinds of ClassLike
operations with fields, that can share functionality with the original
ClassOp. One immediate use case is a class declaration with only types
in its fields, suitable for use to declare external classes.