[Ibis] Boilerplate for a new dialect
We will be porting an internal language over to CIRCT. This dialect will
be the entry point into CIRCT. Eventually, all of the backend will be
part of CIRCT and the frontend parser will be open sourced.

We plan on developing this dialect in the open so as to not dump a giant
PR at some point months down the road.