[Verif] Add `fstr, printf` operations
#5577, #5575

Re. the lowering; `verif.printf` lowers directly to a `sv.fwrite` via. inspection of its source operand (which is expected to be a `verif.fstr`). The source `fstr` is `not` deleted in the process (there may be other users). However, i marked `verif.fstr` as `Pure` meaning that it'll get removed if `canonicalize` is run after `lower-verif-to-sv` and there are no more users of the result.
It's expected that `verif.printf` will reside inside *something* that eventually lowers to a procedural SV region, if one is going the SV route. I'd imagine that most users will embed `verif.printf` inside `hw.triggered` regions.