[FIRRTL][GrandCentral] Stop using ModuleNamespace for unique names (#5619)
ModuleNamespace is used for tracking used and generating new
`inner_sym`s.  This code is mistakenly using it to generate unique
`name`s for wires and ports.  It does not matter if the port has a
unique name, as it will be legalized later in the pipeline.

Additionally, this does not add newly drilled ports as valid targets
for annotations.