[Support] Add target-triple-like option
Starting all the way back in https://discourse.llvm.org/t/target-triples-for-hardware-compilers/4125 and discussed multiple times in the ODM.

Since this is a can of worms, i'm deliberately not calling this "CIRCT target triple".

TargetOptions can be fully generic, and is parsed similar to how a pass pipeline is parsed.

e.g.:
> `target-options='language=sv(standard=1995,no-systemverilog-struct))-target=fpga(family="7-series",model="...")-vendor=mentor'`