[TypeAlias] Add LowerToHW support for lowering TypeAlias (#5292)
Lower `firrtl.alias` to `hw.typealias`. 
Create a global `hw.type_scope` and add the `hw.typedecl`s corresponding to all
the `firrtl.alias`s across  the `firrtl.module`s. 
This PR adds a pre-scan to the `LowerToHW` pass to lower all the `firrtl.alias`
types for all the modules, sequentially. This ensures that name conflict
resolution and declaration order of `typedecl` are deterministic. The
sequential pass records all the alias types and their corresponding lowering,
which is accessed during the parallel phase.