[NFC] Narrow trait used for InstanceGraph
Lots of dialects grab HWModuleLike and don't implement most of it.  Much of use seems to be from wanting to use instance graph.  So this splits out a minimal interface for that use to allow less dependencies on big HW-specific interfaces.

ModuleLike should inherit from Symbol, but this angers table gen.