[circt-reduce] Add DropNames Preserving None Pass
Add the DropNames pass to circt-reduce with the option to drop all names.
This has the benefit of letting canonicalizers/folders go ham.  This was
observed as a missed opportunity on a recent reduction that left many
connections like the following:

    node b = a
    node c = b
    node d = c

Dropping names causes all of these to get folded away and dramatically
reduced the final, already very small, reduced design (by 50%).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>