[NFC] Use ModuleType to compute many properties
In anticipation of most modules converting from functions to portlists using module type, route most of the HWModuleLike interface through the ModuleType.  Modules like things must now be able to produce a moduletype from which many accessors are derived.  getNumPorts is separate from this as non-function-like modules need this to compute the module type.