[WIP, NFC]   Use ModuleType as primary data structure for HW.*Module
This converse HWModules over to using ModuleType to store the port types and names.

This doesn't quite break the assumption that ports are stored in-order.

HWModules are moved away from FunctionOpInterface, necessitating the temproary implementation of some of that functionality.  This is temporary as the separation betwen argument and result in the side-arrays (loc, attrs, etc) will be eliminated in a future patch.