[HW][HWModules] Add ability to set/erase port symbols.
This should be added to an interface shared elsewhere,
(and with different signature suitable to per-field syms)
but for now keep this on the HW module ops.

The `hw.exportPort` string literal is used on HW ops
to store the inner symbol (`getArgSym`, so on).

Also, since the code for HWModule ops in particular already
hardcodes use of `hw.exportPort`, change the getter to
look for the port symbol under that name specifically.