[ci] Add ability to skip CIRCT tests in nightly
Add an environment option to skip CIRCT testing.  This is intended to be
used if the previous job knows that tests don't need to be run (or for
testing of actions with a slightly faster turnaround).

Expose this option as a workflow_dispatch input to make this easier to test.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>