[FIRRTL] Use symbols in ResolveTraces verbatims
Change the ResolveTraces pass to use symbols instead of hard-coded names.
Previously, the pass used hard-coded names derived from the hierpath op
symbols (which was mega-jank! I'm sorry!).  Correct all this by using symbols
with substitutions.

Update the end-to-end test to use a test case which uncovered this
error (because it uses an instance path).  Update the pass unit tests to use the
new format.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>