ExportVerilog emits ports preserving their order. (#5738)
Using the HWTestModuleOP for testing, ExportVerilog is updated to use the port order in the module when emitting instances and modules. HWModuleOP and company are still input,output ordered until their implementation is migrated to the one used in hwtestmoduleop.

The added HWTestModuleOp is temporary to allow development until moving all the modules over.