[FIRRTL] Add ObjectOp (#5759)
ObjectOp is a declaration of an object in firrtl, which in turn is an instance
of a class. ObjectOp has a single return value, of type ClassType, which
represents the instantiated object. Objects are bundle-like, in that they have
input and output ports, but unlike modules, are not exploded into their ports.