[ExportVerilog] Fix Expr/PropertyEmitter token buffer crash
Fix an issue in `ExportVerilog` where `ExprEmitter`s and
`PropertyEmitter`s would not properly use a token buffer passed in as a
constructor argument. The emitters have a local buffer of tokens that is
used by default if no external one is provided. However, parts of the
emitters failed to use `buffer.tokens` to access either the externally
provided or the local buffer, and instead would always use the local
`tokens` buffer. This caused parentheses and end tokens to be lost if an
external buffer was provided, since the emitter would modify the unused
local `tokens` buffer.

This fixes the issue by properly using `buffer.tokens` everywhere and
renaming `tokens` to `localTokens` to make misuse more obvious.

Fixes #5763.