[Ibis] Add instance/port references and related ops
Since port reads/writes are now done at a classref-based level, this also means that there's no reason to have specific instance reads/writes.

A bit of an outstanding question - `ibis.container` was isolated from above, which i still think is the right descision. However, we need to be able to grab `%this` somehow...

Some options:
1. Add a block argument to containers, e.g. `ibis.container @D(%parent)` wherein %parent is implied to be %this of the class
2. Disallow IsolatedFromAbove (what's currently done).