[FIRRTL][FIRParser] Prefer RWProbe op as much as possible.
Keep forceable for now-- it's needed for uninferred resets presently,
and has slightly better (if not ideal) behavior when attempting
to rwprobe a non-passive aggregate.

Note: force_nonpassive.fir doesn't work with symbols, check.

We can't rwprobe() an aggregate that /must/ be lowered.  Forceable
tries to allow this, but when using symbols this is just rejected.

Probably this should not be allowed.