[FIRRTL] Add fir tests relating to probes, from spec.
Examples tweaked as-needed to make full examples,
portion from spec are marked as well as any modifications.

See comments in each, a few are modified to only test
the demonstrated feature or to workaround spec issues
(notably can't express non-const node with literal initializer).

Check these are parsed and make it through the pipeline,
with and without aggregate preservation options.

Test invalid or unsupported examples are appropriately
rejected.

Ensure the examples from spec are under test.