[Pipeline] Remove `ext` input from pipelines
Instead, remove `IsolatedFromAbove` from the pipeline, and define any used value defined outside of the pipeline as an external input. The motiviation for this is to reduce the headache introduced by having to explicitly modify the `ext` input list of a pipeline if more external inputs are to be added into a pipeline (or hoisted out of).