[FIRRTL] Canonicalize Bits(And) to And(Bits) only for single use
Canonicalizations of the for `X(Y(a, b)) -> Y(X(a), X(b))` can be problematic
when the original inner expression has mutliple uses: after pushing through the
canonicalization, the original calculation is still needed. This can sometimes
lead to an explosion of logic in the module, which has real cost in hardware.
This changes the the BitsOfAnd canonicalization to be gated on the inner
expression having a single use.