[GenStateStruct] Initialize a new pass - GenStateStruct pass
This pass collects elements: module ports, output reg of module, local
reg / mem  in current module, signal with attribute of trigger and
instance pointer. Form a struct with these elements and their clone
member which will be used to represent previous one of the member state.
The struct will be stored in the global variable and the pointer of the
struct will be passed to the module. The module will be overrided with
the new struct pointer.