[docs] Add top-level document on hls
This should serve as a landing page for everything HLS-related in CIRCT, given a proper introduction to the various flows, their state, and further pointers on how to find more information.

I'd hope for it to be written with a target audience of "i'm aware of HLS, i'm aware of MLIR, i'm somewhat unaware of CIRCT".